restart-or-new
Restart PuTTY if current window focus is on an Inactive PuTTY Session. Will move mouse to Menu, click, and send "r" key to automate the session restart. Alternatively, this feature exists currently disabled by default to access the menu via Alt+Space or just Alt. But this requires editing currently saved profiles and is still an extra step. Tied to a StreamDeck/Macro button this makes it even more worth it.